fix: compilation errors for Arduino IDE
- parseEventItem: change JsonObject& to const JsonObject& (rvalue binding)
- readKey: keys() → keyList() (newer M5Cardputer API)
- setCursor: max(int16_t, int) → max((int)x, 24) (template deduction fix)

+++ ui.cpp
@@ -84,7 +84,7 @@ void UIManager::drawMessage() {
     M5Cardputer.Display.setTextColor(_msgColor, COLOR_SURFACE);
     M5Cardputer.Display.setTextSize(1);
     int16_t x = 160 - (textWidth(_message) / 2);
-    M5Cardputer.Display.setCursor(max(x, 24), 112);
+    M5Cardputer.Display.setCursor(max((int)x, 24), 112);
     M5Cardputer.Display.print(_message);
 }
